hypertrace: Import the entirety of HyperTrace.
This commit squashes the entire history of HyperTrace into a single commit.
This was done because the original repository was based on the old FreeBSD
GitHub repository that has since been force pushed to due to FreeBSD moving
to git. Instead of keeping that repository, this commit imports the entire
HyperTrace diff into a new 14-CURRENT repository in a single patch.