dtraced, dtrace, bhyve: HyperTrace refactor.
This commit refactors a number of things in dtraced, dtrace and bhyve.
It fixes most of the dtraced deadlocks, crashes and race conditions. It
also implements new functionality in bhyve to listen for dtraced which
has been started after the VM, allowing each VM to be instrumented even
if dtraced gets restarted on the host.

Furthermore, this commit speeds dtraced up by a very large margin due to
simplification of locks, bug fixes and various other things.