dtrace: Ensure that the typefile list is in order.
So far, we were populating the list arbitrarily and simply requiring
that the D CTF container is the very first one. However, a lot of the
code was assuming to check D first and then the kernel. This commit
should make that code a bit happier once migrated to the new API.