dtrace: Potentially address a kernel panic.
The panic happens when we call a hypertrace probe with the vmhdl in the
current thread set, and then a profile probe fires on the host just as
we are enabling interrupts in dtrace_probe() again. The problem is that
the vmhdl is still attached to curthread, but the probe that we are
executing is not a hypertrace probe.

In some conditions, this can cause a kernel panic and this commit
hopefully addresses it.