hypertrace: Fix a whole bunch of races.
This commit fixes a ton of race conditions that existed since the device
driver was written. This was simply forgotten about during the quick
prototyping stage and everything "appeared" to work.

Furthermore, change up some of the return codes and add a sysctl to see
how many probes have been enabled.