dtraced: C++ify client handling.
Transform dtraced_fds into client_fds, a C++ version of a reference
counted file descriptor used internally to know when it's safe to
close() sockets. This commit also restructures some of the internal
functions across files. Since a lot of the functions that operate on
state were sprinkled across many files but now belong in
dtraced_state.cc, this commit begins moving them slowly.

Unfortunately client_fds still need to be pointers because they cannot
live in a single container, and copying the data is not an option as a
stale reference to a client_fd might decrement the reference once it
releases it, but it will only happen in the old copy, not the new one
that we care about. We also have to take special care with client_fds
that originate from an event's udata.