dtraced: Major refactor into C++.
This commit mostly refactors the job struct into a class, but also
moves around a lot of things relating to jobs into the state class.

This is because the state class is responsible for managing its own
data, including a list of jobs, so it should be there instead of
bundled with job-related things like it was in the C version.