dtrace: Use C++ in libdtrace.
This commit rewrites the HyperTrace bits of libdtrace in C++ in order to
make use of STL containers. This greatly simplified the code.

This is only an initial port to C++ rather than something that intends
to be idiomatic C++. That will come in the later commits.

Currently, WITH_DTRACE_DEBUG_ASAN= is broken.