dtrace: Correctly check for D CTF container.
When we compare against the D CTF container in the HyperTrace linker, we
can't just compare the pointer values. This was a leftover from a
work-in-progress implementation. We need to carry through the container
name, which we can then strcmp(). This still doesn't work because we
need to actually create this type in the D container on the new DTrace
instance, so we will need to serialize the D container types to be
created. However, this is the first step in getting it to work.